requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
django-allauth==0.48.0
django-autoslug==1.9.8
django-cities-light==3.9.2
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-filter==2.4.0
django-modeltranslation==0.18.4
django-nine==0.2.5
django-oauth-toolkit==1.7.0
dj-rest-auth==2.2.3
django-rest-swagger==2.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
drf-yasg2==1.19.4
cryptography==36.0.1
progressbar2==4.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.3
python-decouple==3.6
